AG-16452 Split data-selection-zoom benchmark by series family
The combined example ran 24 test cases over a 1M-point dataset in one page
session, accumulating enough memory pressure to dominate benchmark CI
wall-clock (~18 min per shard on hosted runners). Each series family now
runs as its own example with a fresh page session:

- data-selection-zoom-line-area: line, area, stacked-area, range-area
- data-selection-zoom-bar: bar, stacked-bar, range-bar
- data-selection-zoom-financial: candlestick, ohlc
- data-selection-zoom-xy: scatter, bubble, histogram (number axes, xy zoom)

Test cases and measurements are unchanged. The drag-select/click helpers
move to the shared benchmarkUtils.ts. Sharding returns to plain round-robin
since no example needs an isolated shard any more.

Also fixes the ag-shared precommit format gate to skip symlinks — their
staged blob is the target path, which prettier mangles (SYNC-LOG entry
added).
