+++ external/ag-shared/docs/SYNC-LOG.md
@@ -6,6 +6,21 @@ Newest entries first. Generated by `/ag-shared-sync-log`.
 
 ---
 
+## 2026-06-12 -- Commit gate: skip symlinks
+
+**Branch:** `ag-16452/charts-switch-to-browser-based-benchmark`
+**Ticket:** AG-16452
+
+### Changes
+
+- **[scripts/claude-hooks/precommit-check.sh]** Skip symlinks (staged mode `120000`) in the formatting check. A symlink's staged blob is its target path, not file content, so piping it through prettier always produced a spurious "not formatted" denial — e.g. the symlinked `benchmarkUtils.ts` in benchmark example folders.
+
+### Migration Actions
+
+- [ ] None — script is consumed in place; takes effect on next sync.
+
+---
+
 ## 2026-06-10 -- Batched Claude formatting hooks (track / Stop-format / commit-gate)
 
 **Branch:** `ajt/claude-format-hooks`
