+++ packages/ag-charts-website/src/content/docs/benchmarks/_examples/data-selection-zoom-bar/benchmarkUtils.ts
