+++ packages/ag-charts-website/src/content/docs/benchmarks/_examples/data-selection-zoom-bar/data.ts
