+++ packages/ag-charts-website/src/content/docs/benchmarks/_examples/data-selection-zoom-financial/benchmarkHarness.ts
@@ -1,0 +1,1 @@
+../../benchmarkHarness.ts
