+++ packages/ag-charts-website/src/content/docs/benchmarks/_examples/data-selection-zoom-financial/index.html
@@ -1,0 +1,13 @@
+<div class="example-controls">
+    <div class="controls-row">
+        <span>Series Type:</span>
+        <select onchange="setSeries(event.target.value)">
+            <option value="candlestick" selected>Candlestick</option>
+            <option value="ohlc">OHLC</option>
+        </select>
+        <span class="gap-left">Selection:</span>
+        <button onclick="selectClear()">Clear</button>
+        <button onclick="selectHalf()">50% (drag left half)</button>
+    </div>
+</div>
+<div id="myChart"></div>
