+++ packages/ag-charts-website/src/content/docs/benchmarks/_examples/data-selection-zoom-line-area/benchmarkUtils.ts
@@ -1,0 +1,1 @@
+../../benchmarkUtils.ts
