+++ packages/ag-charts-website/src/content/docs/benchmarks/_examples/data-selection-zoom-line-area/index.html
@@ -4,16 +4,8 @@
         <select onchange="setSeries(event.target.value)">
             <option value="line" selected>Line</option>
             <option value="area">Area</option>
-            <option value="bar">Bar</option>
-            <option value="stacked-bar">Stacked Bar</option>
             <option value="stacked-area">Stacked Area</option>
             <option value="range-area">Range Area</option>
-            <option value="range-bar">Range Bar</option>
-            <option value="candlestick">Candlestick</option>
-            <option value="ohlc">OHLC</option>
-            <option value="scatter">Scatter</option>
-            <option value="bubble">Bubble</option>
-            <option value="histogram">Histogram</option>
         </select>
         <span class="gap-left">Selection:</span>
         <button onclick="selectClear()">Clear</button>
